back-end/requirements.txt
easyocr==1.7.1
opencv-python==4.10.0.84
pdf2image==1.17.0
Pillow==10.4.0
numpy==2.1.2
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
groq==0.30.0
httpx==0.27.2  # Downgraded to avoid proxies TypeError with groq SDK
requests==2.32.4
pytest==8.3.3
SQLAlchemy==2.0.35
psycopg[binary]==3.3.1
